Added checks to renaming thread folders. Renamed General to Utils.
Added checks to renaming thread folders, checks if the thread folder already exists, if it does not then it aborts. If it exists and the destination exists also, it will increment an integer in brackets appended to the destination folder ie $"{path} (3)" until a free folder name is found.
Renamed General to Utils short for Utility/Utilities.