Large refactor of "Imageboard" code, rework of thread subject handling.
Previously tracked threads and boards were held under a single "Imageboard" class that could either be tracking a board or a thread, determined by a boolean called isBoard. This has been split into 3 classes, initially Tracker, which is then extended by either Board or Thread, which is then extended by specialised classes like Thread_4Chan or Board_8kun, providing the code needed for scraping these different sites.
Thread subjects have been reworked.
There is now a single "subject" variable, it can either be null, meaning no subject has been grabbed or set, or set.
The thread's actual subject is stored in this variable, and can be overwritten by the user. Even if it is not overwritten (is the original thread name), it is now saved in the threads.dat file. 
Meaning ALL threads in threads.dat now have a subject parameter appended on the end.
Now when opening the program there is no longer a need to re-grab the thread's subject from the net, it is just loaded in from the file. Making opening the program now lightning-fast.
Now the only time the subject will be grabbed from the net now is when the thread is added to tracking.
Changed opening the program from notification tray to a single click instead of double click.
8kun code refactored but still work in progress.