Scan optimisations. Threads now save the greatest Tim downloaded, and skip checking all files with a Tim lower than the saved greatest Tim. Logging Improvements.
All files on 4chan are saved with a "Tim", this is a unix timestamp with microseconds (e.g. 145251205.jpg).
Each thread now stores in a long `GreatestSavedFileTim` which is a record of the file saved with the highest Tim. From now on only files with Tims greater than the saved one will be attempted to be saved.
This means all `File.Exists` checks are now gone, and that when the filename format is changed already saved files will not be updated.

CODE CHANGES:
Added `GreatestSavedFileTim` field to `Thread` class.
Added `Tim` field to `ImageLink` class.
Tims are now saved with thread urls in threads.dat file delimited by a space character (' ').
These tims are now also accounted for in `Utils.CreateNewTracker()`.
`Thread_4Chan.Download()` now checks that a file's Tim is greater before saving it.

Logging improvements:
Program.Log now prints to console if in DEBUG mode.
DEBUG folder in `Application.CommonAppDataPath` is now made (wasn't being made and was causing exceptions).