Adding 8kun support. Now working 100%, but some woes. Listed below:
File tims are a long string of letters and numbers, and they do not seem to "go up" (the tim of a file posted later can be lower or greater than a previous one  (tested using String.Compare)). To combat this the reply's no (number) is used as the file's tim in the ImageLink objects created. This works fine but there might be a better way.
Big woe: About 50-60% of files return 404 from https://8kun.top/file_store/id.jpg and are instead located at 'https://media.8kun.top/board/src/id.jpg '. At the moment at download is attempted, and if it errors then the other url is used. There must be a way to know about this beforehand (causing exceptions is bad).