GridView now updates Subject&FileCounts without need for refreshing. File Counts now display "-" until a count is retrieved. GridView scroll bars not appearing issue resolved. Loading threads greatly optimised by adding threads to list on UI thread.Code neatened.
Thread loading is greatly improved at launch time by skipping all of the pre/post-amble of fiddling with the gridview's scrollbars and columnresize properties and just adding threads on the UI thread. This also fixes the issue of the scrollbar not appearing sometimes.
Threads now update of their own accord when the FileCount or Subject is changed. 
Thread's FileCount now displays a hyphen ("-") up until their first scrape and a filecount is retrieved. This is indicated in code with a constant value of -1.
Removed NotifyPropertyChanged code in Threads_ListChanged event in MainFormModel.cs. This was causing the GridView to scroll back up to the top when done. And when saving the topRowIndex and re-applying it after NotifyPropertyChanged was stopping manual scrolling performed with the mouse  dragging the scroll bar, leading to an unpleasant user experience. This should also lead to some better performance.